+++ lib/shipit.rb
@@ -57,6 +57,7 @@ module Shipit
   delegate :table_name_prefix, to: :secrets
 
   attr_accessor :disable_api_authentication
+  attr_writer :automatically_prepend_bundle_exec
 
   def app_name
     @app_name ||= secrets.app_name || Rails.application.class.name.split(':').first || 'Shipit'
@@ -204,6 +205,17 @@ module Shipit
 
   def default_inactivity_timeout
     secrets.commands_inactivity_timeout || 5.minutes.to_i
+  end
+
+  def automatically_prepend_bundle_exec
+    unless defined?(@automatically_prepend_bundle_exec)
+      ActiveSupport::Deprecation.warn(
+        'Automatically prepending `bundle exec` will be removed in a future version of Shipit, '\
+        'set `Shipit.automatically_prepend_bundle_exec = false` to test the new behaviour.',
+      )
+      @automatically_prepend_bundle_exec = true
+    end
+    @automatically_prepend_bundle_exec
   end
 
   protected